refactor: extract logutil package with concrete Logger and ColorizeYAML
