+++ cmd/init.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 
 	"github.com/rockholla/gitspork/internal"
+	"github.com/rockholla/gitspork/internal/logutil"
 	"github.com/spf13/cobra"
 )
 
@@ -35,7 +36,7 @@ func (isc *InitSubcommand) GetCmd() *cobra.Command {
 	var cmd = &cobra.Command{
 		Use:   "init",
 		Short: initHelpShort,
-		Long:  fmt.Sprintf("%s\n\n%s", initHelpShort, fmt.Sprintf(initHelpLongTemplate, internal.ColorizeYAML(configSchema), internal.ColorizeYAML(migrationsSchema))),
+		Long:  fmt.Sprintf("%s\n\n%s", initHelpShort, fmt.Sprintf(initHelpLongTemplate, logutil.ColorizeYAML(configSchema), logutil.ColorizeYAML(migrationsSchema))),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			return internal.Init(initPath, logger)
 		},
