+++ internal/gitspork_test.go
@@ -9,6 +9,7 @@ import (
 	gogit "github.com/go-git/go-git/v6"
 	"github.com/go-git/go-git/v6/plumbing"
 	"github.com/go-git/go-git/v6/plumbing/object"
+	"github.com/rockholla/gitspork/internal/logutil"
 	"github.com/rockholla/gitspork/internal/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -27,7 +28,7 @@ func TestIntegrate(t *testing.T) {
 		makeUpstreamRepo(t, upstreamDir)
 
 		_, err = Integrate(&types.IntegrateOptions{
-			Logger:              NewLogger(),
+			Logger:              logutil.New(),
 			UpstreamRepoURL:     upstreamDir,
 			UpstreamRepoVersion: "master",
 			DownstreamRepoPath:  downstreamDir,
