+++ internal/init_test.go
@@ -5,6 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
+	"github.com/rockholla/gitspork/internal/logutil"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -12,7 +13,7 @@ func TestInit(t *testing.T) {
 	initPath, err := os.MkdirTemp("", "gitspork-tests")
 	defer os.RemoveAll(initPath)
 	assert.Nil(t, err)
-	err = Init(initPath, NewLogger())
+	err = Init(initPath, logutil.New())
 	assert.Nil(t, err)
 	initedConfigBytes, err := os.ReadFile(filepath.Join(initPath, gitSporkConfigFileName))
 	assert.Nil(t, err)
